Base Setup Part 1
Setting up the navbar on each page, and creating the respective page for each grid that's being worked on. Currently uploading this because Live Server is not loading in the image properly for the background on the home page and need to check if there's issues with it connecting.

All active links for each grid on the navbar were changed properly.

CSS file has all the proper coding in it. Restarting VSCode to attempt fixing the issue.